refactor(patrol): share legacy state persistence
