+++ wiki/testing.md
@@ -114,7 +114,7 @@ task default: :test
 | `web/agents_auth_test.rb`, `web/agents_auth_login_test.rb`, `web/agents_routes_test.rb` | `Hive::Web::AgentsAuth` — Claude paste-back PTY login URL capture, Codex and Grok `--device-auth` poll-login behavior, `gh auth login --web` URL capture plus auto-Enter prompt handling, binary PTY output scrubbing, rejected-code errors, watchdog/process-group cleanup, concurrent-session cap, Pi token JSON rejection/persistence, and route wiring. |
 | `web/config_test.rb`, `web/supervisor_test.rb`, `web/app_coverage_test.rb` | Hivebox config/supervisor packaging support — global web defaults/validation, child restart/backoff/reload/shutdown decisions, and route coverage attribution guardrails. |
 | `patrol/pr_opener_test.rb`, `patrol/review_handoff_test.rb` | `Hive::Patrol::PrOpener` / `ReviewHandoff` — exact base/head and clean-worktree binding, leased/verified remote publication, created/existing PR reconciliation, final remote base/head recheck before task publication, stale-base failed-handoff retry refusal, durable `reconciliation_pending` URL/patch receipts across lookup/auth failures, fail-closed exact-diff/title/body secret scanning, closed results, proof handoff, and exact-patch handoff retry. |
-| `patrol/feature_batch_test.rb`, `patrol/candidate_selector_test.rb`, `patrol/reviewer_test.rb`, `patrol/fingerprint_test.rb`, `patrol/fixer_test.rb` | Ordinary patrol alpha policy — exact detached SHA-bound review rotation across default advances and first-batch errors, observable partial reviews, bounded/atomic source-matching evidence, immutable finding ids, legacy/current and same-run deduplication, retryable pending publication, model-label-light alpha ranking, strict fresh-base creation with no stale-local fallback, language-neutral declared regression paths, guardrails, nonblocking no-follow regular-file proof reads, normal regression-identified fail-before/pass-after proof, and non-suppressing fixer rejection. |
+| `patrol/value_objects_test.rb`, `patrol/feature_batch_test.rb`, `patrol/candidate_selector_test.rb`, `patrol/reviewer_test.rb`, `patrol/fingerprint_test.rb`, `patrol/fixer_test.rb` | Ordinary patrol alpha policy and shared legacy state persistence — tolerant hash-only JSON reads, exact detached SHA-bound review rotation across default advances and first-batch errors, observable partial reviews, bounded/atomic source-matching evidence, immutable finding ids, legacy/current and same-run deduplication, retryable pending publication, model-label-light alpha ranking, strict fresh-base creation with no stale-local fallback, language-neutral declared regression paths, guardrails, nonblocking no-follow regular-file proof reads, normal regression-identified fail-before/pass-after proof, and non-suppressing fixer rejection. |
 | `patrol/source_reader_test.rb`, `patrol/mapper_test.rb`, `patrol/architecture_mapper_test.rb`, `refactor_patrol/leverage_test.rb` | Architecture source inspection — polyglot/unknown-text/shebang and bounded docs mapping, source-only dependency leverage, canonical-root confinement at both the outer semantic mapper and architecture mapper, external package/Python/docs symlink and device rejection before feature construction, confined in-repository symlink support, regular-file checks, UTF-8 scrubbing, and the 256 KiB per-source read cap. |
 | `refactor_patrol/*_test.rb` | Architecture patrol — strict run-wide reviewer output and wall-clock budgets, root-confined and 256 KiB-bounded real-byte evidence verification (including sparse oversize sources and beyond-cap rejection), immutable manifests/policy, exhaustive dispositions and partial feature resume, host-namespaced semantic families/canonical actions, exact-marker-first legacy issue-family reconciliation, immutable global proof archives plus disposable catalog rebuild (including handoff recovery after a synthetic task advances through coding's later stages), exact-registration ownership including disabled pending continuations, repository-global fix branches, confined workspace-write fixing and public-contract guards, verified Claude safe-mode discovery, generation-fenced job/action transitions, PR/issue intent and reconciliation, OPEN-draft rejection, created-PR head/base OID verification, fenced mandatory review handoff, fully non-mutating dry-run parity, and mixed fix/issue/suppression completion. |
 | `refactor_patrol/job_store_test.rb`, `refactor_patrol/pr_opener_test.rb`, `refactor_patrol/action_runner_test.rb`, `refactor_patrol/repository_ownership_test.rb` | Generation-scoped publication recovery — NUL-delimited base/commit attempt ids, append-only phase/descriptor validation, additive exact-payload legacy-receipt import, pre-create trunk drift rechecks, immutable supersession, exact proven-old-OID branch replacement, and publication phases as continuation ownership. The end-to-end regression uses a real git repository/worktree, bare Git remote, `JobStore`, `ActionRunner`, and `PrOpener`: a forked worker is hard-killed after the remote push but before its completion receipt, trunk advances, a new process owner resolves the expired dead claim and receipts/supersedes the landed patch, then the fixer replaces only the proven old remote OID and creates/verifies/hands off exactly one PR. |
